Up-to-date 2/23

+++ src-test/downloadCenter/TestPlateInfoDownloadStore.java
@@ -13,7 +13,7 @@ public class TestPlateInfoDownloadStore {
 	
 	@Test
 	public void testPlateInfoDownloadStore() {
-		double mockMJD = 1.0, mockPlate = 1.0, mockFiber = 1.0;
+		int mockMJD = 1, mockPlate = 1, mockFiber = 1;
 		DownloadStore store = new PlateInfoDownloadStore(mockMJD, mockPlate, mockFiber, WorkingDirectory.STUB);
 		String formattedForWGET = URL + mockPlate + "/1d/spSpec-" + mockMJD + "-" + mockPlate + "-" + mockFiber + ".fit";
